pyproject.toml
[tool.poetry]
name = "django_project"
version = "0.1.0"
description = ""
authors = ["kot <bafomet2016vitt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.4"
psycopg2 = "^2.9.7"
Pillow = "^10.0.0"
ipython = "^8.14.0"
python-slugify = "^8.0.1"
redis = "^5.0.0"
python-dotenv = "^1.0.0"
django-crontab = "^0.7.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
